requirements.txt
faiss_cpu==1.8.0.post1
Jinja2==3.1.4
openai==1.54.3
python-dotenv==1.0.1
streamlit==1.39.0
utils==1.0.2
